feat(#2148): Remove SkipsTest
